requirements.txt
google-adk
google-generativeai
python-dotenv
litellm
pydantic
google-cloud-aiplatform[adk,agent_engines]
